legal compliance

+++ cleanup.js
@@ -17,6 +17,7 @@ module.exports = {
       var currTime = parseInt(Date.now() / 1000);
 
       if (/fallback/.test(file)) return; // never delete fallback folder(s)
+      if (/keep/.test(file)) return;
       
       if ((currTime - fileTime) > 1800) {
         rmdir(file, () => console.log(`Cleanup: deleted ${file} by experation. ${currTime - fileTime}`))
